// src/api-spec/protobuf/gen/js/tdex/v1/swap_pb.d.ts
export interface SwapRequest {
  /**
   * Random unique identifier for the current message
   *
   * @generated from protobuf field: string id = 1;
   */
  id: string;
  /**
   * The proposer's quantity
   *
   * @generated from protobuf field: uint64 amount_p = 2;
   */
  amountP: string;
  /**
   * The proposer's asset hash
   *
   * @generated from protobuf field: string asset_p = 3;
   */
  assetP: string;
  /**
   * The responder's quantity
   *
   * @generated from protobuf field: uint64 amount_r = 4;
   */
  amountR: string;
  /**
   * The responder's asset hash
   *
   * @generated from protobuf field: string asset_r = 5;
   */
  assetR: string;
  /**
   * The proposer's unsigned transaction in PSBT format (base64 string)
   *
   * @generated from protobuf field: string transaction = 6;
   */
  transaction: string;
  /**
   * In case of a confidential psetv0 transaction the blinding key of each
   * confidential input is included. Each blinding key is identified by the
   * prevout script hex encoded.
   *
   * @generated from protobuf field: map<string, bytes> input_blinding_key = 7;
   */
  inputBlindingKey: {
    [key: string]: Uint8Array;
  };
  /**
   * In case of a confidential psetv0 transaction the blinding key of each
   * confidential output is included. Each blinding key is identified by the
   * output script hex encoded.
   *
   * @generated from protobuf field: map<string, bytes> output_blinding_key = 8;
   */
  outputBlindingKey: {
    [key: string]: Uint8Array;
  };
}
/**
 * @generated from protobuf message tdex.v1.SwapAccept
 */
export interface SwapAccept {
  /**
   * Random unique identifier for the current message
   *
   * @generated from protobuf field: string id = 1;
   */
  id: string;
  /**
   * indetifier of the SwapRequest message
   *
   * @generated from protobuf field: string request_id = 2;
   */
  requestId: string;
  /**
   * The partial signed transaction base64 encoded containing the Responder's
   * signed inputs in a PSBT format
   *
   * @generated from protobuf field: string transaction = 3;
   */
  transaction: string;
  /**
   * In case of a confidential transaction the blinding key of each confidential
   * input is included. Each blinding key is identified by the prevout script
   * hex encoded.
   *
   * @generated from protobuf field: map<string, bytes> input_blinding_key = 4;
   */
  inputBlindingKey: {
    [key: string]: Uint8Array;
  };
  /**
   * In case of a confidential transaction the blinding key of each confidential
   * output is included. Each blinding key is identified by the output script
   * hex encoded.
   *
   * @generated from protobuf field: map<string, bytes> output_blinding_key = 5;
   */
  outputBlindingKey: {
    [key: string]: Uint8Array;
  };
}
/**
 * @generated from protobuf message tdex.v1.SwapComplete
 */
export interface SwapComplete {
  /**
   * Random unique identifier for the current message
   *
   * @generated from protobuf field: string id = 1;
   */
  id: string;
  /**
   * indetifier of the SwapAccept message
   *
   * @generated from protobuf field: string accept_id = 2;
   */
  acceptId: string;
  /**
   * The signed transaction base64 encoded containing the Proposers's signed
   * inputs in a PSBT format
   *
   * @generated from protobuf field: string transaction = 3;
   */
  transaction: string;
}
// ...
declare class SwapRequest$Type extends MessageType<SwapRequest> {
  constructor();
  create(value?: PartialMessage<SwapRequest>): SwapRequest;
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: SwapRequest
  ): SwapRequest;
  private binaryReadMap7;
  private binaryReadMap8;
  internalBinaryWrite(message: SwapRequest, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter;
}
/**
 * @generated MessageType for protobuf message tdex.v1.SwapRequest
 */
export declare const SwapRequest: SwapRequest$Type;
declare class SwapAccept$Type extends MessageType<SwapAccept> {
  constructor();
  create(value?: PartialMessage<SwapAccept>): SwapAccept;
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: SwapAccept
  ): SwapAccept;
  private binaryReadMap4;
  private binaryReadMap5;
  internalBinaryWrite(message: SwapAccept, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter;
}
/**
 * @generated MessageType for protobuf message tdex.v1.SwapAccept
 */
export declare const SwapAccept: SwapAccept$Type;
declare class SwapComplete$Type extends MessageType<SwapComplete> {
  constructor();
  create(value?: PartialMessage<SwapComplete>): SwapComplete;
  internalBinaryRead(
    reader: IBinaryReader,
    length: number,
    options: BinaryReadOptions,
    target?: SwapComplete
  ): SwapComplete;
  internalBinaryWrite(message: SwapComplete, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter;
}
/**
 * @generated MessageType for protobuf message tdex.v1.SwapComplete
 */
export declare const SwapComplete: SwapComplete$Type;